packages/python-core/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "versapy"
version = "1.1.0"
description = "Lightweight framework for desktop apps with web UI."
authors = [
  { name = "lso", email = "lso.contact@proton.me" }
]
readme = "README.md"

license = "MIT"
requires-python = ">=3.9"
dependencies = [
  "python-socketio", 
  "pywebview", 
  "pyinstaller", 
  "uvicorn", 
  "fastapi",
  "pydantic"
]

[tool.setuptools.packages.find]
where = ["."]
include = [
  "versapy", "versapy.*"
]
